pyproject.toml
[project]
name = "notification-gitlab-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.20.0.post0",
    "python-gitlab>=5.6.0",
    "pyyaml>=6.0.2",
    "sqlalchemy>=2.0.41",
]
[tool.uv]
dev-dependencies = ["ruff>=0.11.7"]

[tool.ruff.lint]
select = ["E", "F", "B"]
ignore = ["E501"]
